change ParsingExpressions to public class
