+++ src/org/peg4d/GrammarFormatter.java
@@ -2,10 +2,11 @@ package org.peg4d;
 
 import java.util.HashMap;
 
+import org.peg4d.expression.*;
 import org.peg4d.vm.Instruction2;
 import org.peg4d.vm.Opcode2;
 
-class GrammarFormatter extends ParsingExpressionVisitor {
+public class GrammarFormatter extends ParsingExpressionVisitor {
 	protected StringBuilder sb = null;
 	public GrammarFormatter() {
 		this.sb = null;
