+++ src/org/peg4d/Main.java
@@ -9,6 +9,7 @@ import java.io.InputStreamReader;
 import java.util.TreeMap;
 
 import org.peg4d.data.RelationBuilder;
+import org.peg4d.expression.ParsingExpression;
 import org.peg4d.ext.Generator;
 
 public class Main {
