+++ src/org/peg4d/Optimizer2.java
@@ -1,5 +1,7 @@
 package org.peg4d;
 
+import org.peg4d.expression.*;
+
 
 
 class Optimizer2 {
