+++ src/org/peg4d/ParsingUtils.java
@@ -1,5 +1,7 @@
 package org.peg4d;
 
+import org.peg4d.expression.ParsingExpression;
+
 public class ParsingUtils {
 	
 	public final static long objectId(long pos, short pegid) {
