+++ src/org/peg4d/expression/ParsingAnd.java
@@ -1,0 +1,34 @@
+package org.peg4d.expression;
+
+import java.util.TreeMap;
+
+import org.peg4d.ParsingContext;
+import org.peg4d.ParsingExpressionVisitor;
+
+public class ParsingAnd extends ParsingUnary {
+	ParsingAnd(ParsingExpression e) {
+		super(e);
+	}
+	@Override ParsingExpression uniquefyImpl() { 
+		return ParsingExpression.uniqueExpression("&\b" + this.uniqueKey(), this);
+	}
+	@Override
+	public ParsingExpression normalizeImpl(boolean lexOnly, TreeMap<String,String> withoutMap) {
+		ParsingExpression e = inner.normalizeImpl(lexOnly, withoutMap);
+		if(e == inner) {
+			return this;
+		}
+		return ParsingExpression.newAnd(e);
+	}
+	@Override
+	public void visit(ParsingExpressionVisitor visitor) {
+		visitor.visitAnd(this);
+	}
+	@Override
+	public boolean simpleMatch(ParsingContext context) {
+		long pos = context.getPosition();
+		this.inner.matcher.simpleMatch(context);
+		context.rollback(pos);
+		return !context.isFailure();
+	}
+}
