+++ src/org/peg4d/expression/ParsingExport.java
@@ -1,0 +1,28 @@
+package org.peg4d.expression;
+
+import java.util.TreeMap;
+
+import org.peg4d.ParsingContext;
+import org.peg4d.ParsingExpressionVisitor;
+
+public class ParsingExport extends ParsingUnary {
+	ParsingExport(ParsingExpression e) {
+		super(e);
+	}
+	@Override
+	ParsingExpression uniquefyImpl() {
+		return new ParsingExport(inner);
+	}
+	@Override
+	public ParsingExpression normalizeImpl(boolean lexOnly, TreeMap<String,String> withoutMap) {
+		return inner.normalizeImpl(lexOnly, withoutMap);
+	}
+	@Override
+	public void visit(ParsingExpressionVisitor visitor) {
+		visitor.visitExport(this);
+	}
+	@Override
+	public boolean simpleMatch(ParsingContext context) {
+		return true;
+	}
+}
