+++ src/org/peg4d/expression/ParsingFunction.java
@@ -1,0 +1,27 @@
+package org.peg4d.expression;
+
+import java.util.TreeMap;
+
+import org.peg4d.ParsingExpressionVisitor;
+
+public abstract class ParsingFunction extends ParsingExpression {
+	public String funcName;
+	ParsingFunction(String funcName) {
+		super();
+		this.funcName = funcName;
+	}
+	@Override
+	public ParsingExpression normalizeImpl(boolean lexOnly, TreeMap<String, String> withoutMap) {
+		return this;
+	}
+	public String getParameters() {
+		return "";
+	}
+	@Override final ParsingExpression uniquefyImpl() {
+		return 	ParsingExpression.uniqueExpression("<"+this.funcName+this.getParameters(), this);
+	}
+	@Override
+	public void visit(ParsingExpressionVisitor visitor) {
+		visitor.visitParsingFunction(this);
+	}
+}
