+++ src/org/peg4d/expression/ParsingList.java
@@ -1,0 +1,59 @@
+package org.peg4d.expression;
+
+import org.peg4d.UList;
+
+public abstract class ParsingList extends ParsingExpression {
+	UList<ParsingExpression> inners;
+	ParsingList(UList<ParsingExpression> list) {
+		super();
+		this.inners = list;
+	}
+	@Override
+	public final int size() {
+		return this.inners.size();
+	}
+	@Override
+	public final ParsingExpression get(int index) {
+		return this.inners.ArrayValues[index];
+	}
+	final void set(int index, ParsingExpression e) {
+		this.inners.ArrayValues[index] = e;
+	}
+	protected final String uniqueKey() {
+		StringBuilder sb = new StringBuilder();
+		for(int i = 0; i < this.size(); i++) {
+			ParsingExpression e = this.get(i).uniquefy();
+			set(i, e);
+			sb.append(e.uniqueId);
+			sb.append(":");
+		}
+		return sb.toString();
+	}
+
+	@Override
+	boolean hasObjectOperation() {
+		for(int i = 0; i < this.size(); i++) {
+			if(this.get(i).hasObjectOperation()) {
+				return true;
+			}
+		}
+		return false;
+	}
+
+	@Override
+	public short acceptByte(int ch) {
+		for(int i = 0; i < this.size(); i++) {
+			short r = this.get(i).acceptByte(ch);
+			if(r != CheckNextFlow) {
+				return r;
+			}
+		}
+		return CheckNextFlow;
+	}
+	
+	public final void swap(int i, int j) {
+		ParsingExpression e = this.inners.ArrayValues[i];
+		this.inners.ArrayValues[i] = this.inners.ArrayValues[j];
+		this.inners.ArrayValues[j] = e;
+	}
+}
