+++ src/org/peg4d/expression/ParsingNot.java
@@ -1,0 +1,58 @@
+package org.peg4d.expression;
+
+import java.util.TreeMap;
+
+import org.peg4d.ParsingContext;
+import org.peg4d.ParsingExpressionVisitor;
+import org.peg4d.ParsingObject;
+
+public class ParsingNot extends ParsingUnary {
+	ParsingNot(ParsingExpression e) {
+		super(e);
+	}
+	@Override ParsingExpression uniquefyImpl() { 
+		return ParsingExpression.uniqueExpression("!\b" + this.uniqueKey(), this);
+	}
+	@Override
+	public ParsingExpression normalizeImpl(boolean lexOnly, TreeMap<String,String> withoutMap) {
+		ParsingExpression e = inner.normalizeImpl(true, withoutMap);
+		if(e == inner) {
+			return this;
+		}
+		return ParsingExpression.newNot(e);
+	}
+	@Override
+	public void visit(ParsingExpressionVisitor visitor) {
+		visitor.visitNot(this);
+	}
+	@Override
+	public short acceptByte(int ch) {
+		short r = this.inner.acceptByte(ch);
+		if(r == Accept) {
+			return Reject;
+		}
+		if(r == Reject) {
+			return Accept;
+		}
+		return r;
+	}
+	@Override
+	public boolean simpleMatch(ParsingContext context) {
+		long pos = context.getPosition();
+		long f   = context.rememberFailure();
+		ParsingObject left = context.left;
+		if(this.inner.matcher.simpleMatch(context)) {
+			context.rollback(pos);
+			context.failure(this);
+			left = null;
+			return false;
+		}
+		else {
+			context.rollback(pos);
+			context.forgetFailure(f);
+			context.left = left;
+			left = null;
+			return true;
+		}
+	}
+}
