+++ src/org/peg4d/expression/ParsingOperation.java
@@ -1,0 +1,26 @@
+package org.peg4d.expression;
+
+import org.peg4d.ParsingExpressionVisitor;
+
+public abstract class ParsingOperation extends ParsingUnary {
+	public String funcName;
+	ParsingOperation(String funcName, ParsingExpression inner) {
+		super(inner);
+		this.funcName = funcName;
+		this.inner = inner;
+	}
+	@Override final ParsingExpression uniquefyImpl() {
+		return 	ParsingExpression.uniqueExpression("<"+this.funcName+this.getParameters()+"+"+this.uniqueKey(), this);
+	}
+	public String getParameters() {
+		return "";
+	}
+	@Override
+	public void visit(ParsingExpressionVisitor visitor) {
+		visitor.visitParsingOperation(this);
+	}
+	@Override
+	public ParsingExpression getExpression() {
+		return this.inner;
+	}
+}
